@@ -1,6 +1,5 @@
-package com.redhat.depdraw.dataservice.dao.model;
+package com.redhat.depdraw.model;
 
-import java.util.List;
 import java.util.Set;
 
 import lombok.AllArgsConstructor;
@@ -17,10 +16,14 @@ import lombok.ToString;
 @ToString(exclude = "uuid")
 @EqualsAndHashCode(exclude = "uuid")
 public class LineCatalog {
+    public static final String INHERIT_LABELS = "Inherit Labels";
+    public static final String INHERIT_ANNOTATIONS = "Inherit Annotations";
+    public static final String INHERIT_METADATA = "Inherit Metadata";
+    public static final String SELECT_RESOURCE = "Select Resource";
+
     private String uuid;
 
     private String name;
 
     private Set<String> rules;
-
 }
