@@ -2,7 +2,7 @@ package com.redhat.depdraw.dataservice.dao.api;
 
 import java.util.List;
 
-import com.redhat.depdraw.dataservice.dao.model.ResourceCatalog;
+import com.redhat.depdraw.model.ResourceCatalog;
 
 public interface ResourceCatalogDao {
 
