@@ -11,12 +11,12 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.redhat.depdraw.dataservice.dao.api.DiagramDao;
-import com.redhat.depdraw.dataservice.dao.model.Diagram;
+import com.redhat.depdraw.model.Diagram;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 
 @ApplicationScoped
 public class DiagramDaoImpl implements DiagramDao {
