@@ -1,9 +1,11 @@
 package com.redhat.depdraw.dataservice.dao.file;
 
+import org.eclipse.microprofile.config.ConfigProvider;
+
 import java.io.File;
 
 public class FileUtil {
-    public static final String BASE_FILES_DIR = "/opt/DepDraw/data/";
+    public static final String BASE_FILES_DIR = ConfigProvider.getConfig().getValue("base.file.dir",String.class);
     public static final String DIAGRAM_FILES_DIR = BASE_FILES_DIR + "Diagrams/";
     public static final String DIAGRAM_FILE_NAME = "Diagram.json";
     public static final String DIAGRAM_RESOURCE_FILE_NAME = "DiagramResource.json";
