@@ -7,12 +7,11 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.redhat.depdraw.dataservice.dao.api.ResourceCatalogDao;
-import com.redhat.depdraw.dataservice.dao.model.ResourceCatalog;
+import com.redhat.depdraw.model.ResourceCatalog;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 
 @ApplicationScoped
 public class ResourceCatalogDaoImpl implements ResourceCatalogDao {
