@@ -3,12 +3,11 @@ package com.redhat.depdraw.dataservice.service;
 import java.util.List;
 import java.util.UUID;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-
 import com.redhat.depdraw.dataservice.dao.api.DiagramResourceDao;
-import com.redhat.depdraw.dataservice.dao.model.Diagram;
-import com.redhat.depdraw.dataservice.dao.model.DiagramResource;
+import com.redhat.depdraw.model.Diagram;
+import com.redhat.depdraw.model.DiagramResource;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 
 @ApplicationScoped
 public class DiagramResourceService {
